09-26-2018. ToDo: Fix the decimal reward.

+++ app/src/main/java/com/github/jc/cryptopea/Utils/Constants.java
@@ -26,7 +26,7 @@ public class Constants {
         return new Random().nextInt((max - min) + 1) + min;
     }
 
-    public String currencyFormatter(double val){
+    public String currencyFormatter(float val){
         return String.format(Locale.getDefault(), "%,.2f", val);
     }
 }
